[Storage] Jumbo blob support (#9649)
Bump block size and max upload size from int to long. Bump blob block size to 4000MB.

+++ sdk/storage/azure-storage-file-datalake/src/main/java/com/azure/storage/file/datalake/DataLakeFileClient.java
@@ -61,7 +61,7 @@ public class DataLakeFileClient extends DataLakePathClient {
     /**
      * Indicates the maximum number of bytes that can be sent in a call to upload.
      */
-    private static final int MAX_APPEND_FILE_BYTES = DataLakeFileAsyncClient.MAX_APPEND_FILE_BYTES;
+    private static final long MAX_APPEND_FILE_BYTES = DataLakeFileAsyncClient.MAX_APPEND_FILE_BYTES;
 
     private final ClientLogger logger = new ClientLogger(DataLakeFileClient.class);
